iOS video_capture: start camera in the background.
Camera start is a blocking operation so never a good idea to do on a main
thread, but worse than that is that the guts of WebView appear to be
interacting with capture start in a bad way causing startup to pause for 10s
while a timeout expires.  This change eliminates that 10s delay.

R=noahric@google.com

Review URL: https://webrtc-codereview.appspot.com/10449004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: b64d52c2924ff5f4970c12391d50612e49000c69