Roll chromium_revision 255773:260462
This disables GN use for the moment (Chromium
has disabled it for now but plan to pick up the
work at a later stage). I'm leaving the rest of
the GN stuff in our DEPS since that's how
the Chromium DEPS currently looks like.

Overview of changes in Chrome DEPS:
$ svn diff http://src.chromium.org/chrome/trunk/src/DEPS -r 255773:260462

which can be compared with the output of:
$ svn cat http://webrtc.googlecode.com/svn/trunk/DEPS | grep chromium_deps | sed 's/^ *//' | sort | uniq

in a WebRTC checkout, gives the following relevant changes:
* third_party/android_tools 0582bd:ca3567
* third_party/icu 249466:259309
* third_party/libjpeg_turbo 251747:259851
* third_party/libyuv 979:986
* third_party/nss 254867:259440
* tools/gyp 1860:1880

The following also shows that Clang is upgraded from r198389 to r202554:
$ svn diff http://src.chromium.org/chrome/trunk/src/tools/clang/scripts/update.sh -r 255773:260462

TEST=trybots
BUG=None
R=tommi@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/10679004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: d10bdd3f78ef22a78eca99febd2f127ebe8ca559