Fixing a bug in ACM2 where the output frame energy was incorrectly set
The value of AudioFrame::energy_ must be set to -1 in order to have the
energy calculated later on in the AudioConferenceMixer module. This was
not the case in ACM2, where the value was set to 0 instead. This
resulted in bad audio for multi-party calls (5 or more participants).

Implemented a unit test to verify ACM output frame.

BUG=3255
TBR=turaj@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/12369005

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 0a2277448ec22b0c9af98d9ebc1acd757be8b89e