Enables VolumeTest.DefaultMicrophoneVolumeIsAtMost255
Rewritten the test to only check for valid volume when we have actually received a value from the audio device. To check if we have actually received a volume value is out of the scope for this test.

BUG=webrtc:367
TESTED=trybots
R=tina.legrand@webrtc.org, xians@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/16499006

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 8d63d0ee70d27439fd153af7093a6be390686efe