Avoid NACK-list flush error on keyframe packets.
Receiver code used to indicate a flush error even if the incoming packet
is a keyframe, forcing a request of a keyframe. Now it takes this
keyframe into account and doesn't error as the stream is decodable from
this point.

BUG=
R=stefan@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/15549005

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: ebb467fdc8d6147ec3e84f7ec9d4336f125687e5