Add missing sources to webrtc/base/base.gyp
During my work on setting up a GN build for
WebRTC, I discovered that the base.gyp is trying
to remove source files (for the Chromium build)
that are not added in the initial set.
I assume these files should be listed and that
GYP just doesn't complain when it's trying to
remove a file that is not present in the sources
list.

natserver_main.cc is also removed, since it's not used anywhere.

There are also a couple of other header files that are
used in other code that probably also should be listed in
base.gyp (please do this in another CL):
* compile_assert.h
* dscp.h
* move.h
* template_util.h

BUG=None
TEST=Trybots passing clobber compile step.
R=henrike@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/13659004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 2bae3211b155d83797150e82da432ec5cf64cff9