Selecting bot_type changed to be specified in the test file
Selecting bot_type changed to be specified in the test file instead of
specify it in the running command.

Now we can write test for rtcBot that run one bot on chrome for android
and the other bot on chrome for desktop.

R=andresp@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/23069004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 3e2f8ff36cbe339ebc35d63d3cf8ed379dda3d59