Revert "Accept incoming pings before remote answer is set to reduce connection latency."
This reverts r7980.

It was causing the ICE connected state to happen while still in the new state rather than going through the checking state, which was causing an ASSERT to fire, which was causing a crash.

Review URL: https://webrtc-codereview.appspot.com/41429004

Cr-Mirrored-From: https://chromium.googlesource.com/external/webrtc
Cr-Mirrored-Commit: 9657265f391cfe473a61b18a4579bbbeb44c9bd8